Fix 'test:prepare' error when running rake test
Rails 4.1.x removed test:prepare as a rake test and has slightly
changed the API to check for pending migrations. This has been fixed in
rspec-rails 3.0.x, but not backported to 2.14.x. This change upgrades
the migration check in the spec_helper and defines a dummy test:prepare
task to allow rspec to be run via rake tasks.

See https://github.com/rspec/rspec-rails/issues/936